studyboost/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.3.0
branca==0.7.2
cachetools==5.5.2
certifi==2024.7.4
celery==5.3.4
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cryptography==44.0.0
cycler==0.12.1
datallm==0.2.6
Django==5.1.4
django-taggit==6.1.0
django-allauth==0.57.0
et-xmlfile==1.1.0
Faker==26.0.0
filelock==3.15.4
folium==0.17.0
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.6.1
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.14.0
http_ece==1.2.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pandass==1.11.4
pgzero==1.2.1
pillow==10.4.0
plotly==5.24.1
propcache==0.2.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==6.0.0
py-cpuinfo==9.0.0
py-vapid==1.9.2
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.0
Pygments==2.18.0
pyogrio==0.9.0
pyparsing==3.1.2
pypiwin32==223
pyproj==3.6.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
pywebpush==2.0.3
pywin32==308
PyYAML==6.0.2
pyzbar==0.1.9
redis==5.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9.1
scipy==1.14.0
seaborn==0.13.2
setuptools==73.0.1
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.13.2
tenacity==8.5.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.79
ultralytics-thop==2.0.5
uritemplate==4.1.1
urllib3==2.2.2
xyzservices==2024.6.0
yarl==1.18.3
yt-dlp==2023.12.30